backend/requirements.txt
# Course Companion FTE - Backend Dependencies
# Generative AI Fundamentals Course

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# R2 Storage
boto3==1.29.0

# CORS
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# Database (for Phase 1 progress tracking)
# Using SQLite for development, can switch to PostgreSQL/Supabase for production
# sqlalchemy==2.0.23  # Uncomment for Phase 1 progress tracking

# Utilities
python-dateutil==2.8.2
